# backend/openapi.yml
openapi: 3.0.3
info:
  title: WhereDidIShoot API
  version: 0.2.0
  description: OpenAPI definition for the Flask backend.
servers:
  - url: /api
    description: Nginx-routed API base path
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      operationId: login
      summary: Login with a username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Existing user logged in successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "201":
          description: Username did not exist and was registered automatically.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users:
    get:
      tags:
        - Users
      operationId: listUsers
      summary: List users
      responses:
        "200":
          description: Users returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRecord"
    post:
      tags:
        - Users
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
  /users/{user_id}:
    parameters:
      - $ref: "#/components/parameters/UserIdPath"
    get:
      tags:
        - Users
      operationId: getUser
      summary: Get a user
      responses:
        "200":
          description: User returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Users
      operationId: updateUser
      summary: Update a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: User updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Users
      operationId: deleteUser
      summary: Delete a user and all child records
      responses:
        "204":
          description: User deleted successfully.
        "404":
          $ref: "#/components/responses/NotFound"
  /practices:
    get:
      tags:
        - Practices
      operationId: listPractices
      summary: List practices
      parameters:
        - $ref: "#/components/parameters/UserIdFilter"
      responses:
        "200":
          description: Practices returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PracticeRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
    post:
      tags:
        - Practices
      operationId: createPractice
      summary: Create a practice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeInput"
      responses:
        "201":
          description: Practice created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
  /practices/{practice_id}:
    parameters:
      - $ref: "#/components/parameters/PracticeIdPath"
    get:
      tags:
        - Practices
      operationId: getPractice
      summary: Get a practice
      responses:
        "200":
          description: Practice returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeRecord"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Practices
      operationId: updatePractice
      summary: Update a practice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeInput"
      responses:
        "200":
          description: Practice updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Practices
      operationId: deletePractice
      summary: Delete a practice and all child records
      responses:
        "204":
          description: Practice deleted successfully.
        "404":
          $ref: "#/components/responses/NotFound"
  /rounds:
    get:
      tags:
        - Rounds
      operationId: listRounds
      summary: List rounds
      parameters:
        - $ref: "#/components/parameters/PracticeIdFilter"
      responses:
        "200":
          description: Rounds returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoundRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
    post:
      tags:
        - Rounds
      operationId: createRound
      summary: Create a round
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoundInput"
      responses:
        "201":
          description: Round created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
  /rounds/{round_id}:
    parameters:
      - $ref: "#/components/parameters/RoundIdPath"
    get:
      tags:
        - Rounds
      operationId: getRound
      summary: Get a round
      responses:
        "200":
          description: Round returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundRecord"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Rounds
      operationId: updateRound
      summary: Update a round
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoundInput"
      responses:
        "200":
          description: Round updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Rounds
      operationId: deleteRound
      summary: Delete a round and all child records
      responses:
        "204":
          description: Round deleted successfully.
        "404":
          $ref: "#/components/responses/NotFound"
  /ends:
    get:
      tags:
        - Ends
      operationId: listEnds
      summary: List ends
      parameters:
        - $ref: "#/components/parameters/RoundIdFilter"
      responses:
        "200":
          description: Ends returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EndRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
    post:
      tags:
        - Ends
      operationId: createEnd
      summary: Create an end
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndInput"
      responses:
        "201":
          description: End created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
  /ends/{end_id}:
    parameters:
      - $ref: "#/components/parameters/EndIdPath"
    get:
      tags:
        - Ends
      operationId: getEnd
      summary: Get an end
      responses:
        "200":
          description: End returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndRecord"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Ends
      operationId: updateEnd
      summary: Update an end
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndInput"
      responses:
        "200":
          description: End updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Ends
      operationId: deleteEnd
      summary: Delete an end and all child records
      responses:
        "204":
          description: End deleted successfully.
        "404":
          $ref: "#/components/responses/NotFound"
  /arrows:
    get:
      tags:
        - Arrows
      operationId: listArrows
      summary: List arrows
      parameters:
        - $ref: "#/components/parameters/EndIdFilter"
      responses:
        "200":
          description: Arrows returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArrowRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
    post:
      tags:
        - Arrows
      operationId: createArrow
      summary: Create an arrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArrowInput"
      responses:
        "201":
          description: Arrow created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArrowRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
  /arrows/{arrow_id}:
    parameters:
      - $ref: "#/components/parameters/ArrowIdPath"
    get:
      tags:
        - Arrows
      operationId: getArrow
      summary: Get an arrow
      responses:
        "200":
          description: Arrow returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArrowRecord"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Arrows
      operationId: updateArrow
      summary: Update an arrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArrowInput"
      responses:
        "200":
          description: Arrow updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArrowRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Arrows
      operationId: deleteArrow
      summary: Delete an arrow
      responses:
        "204":
          description: Arrow deleted successfully.
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    UserIdPath:
      name: user_id
      in: path
      required: true
      description: Numeric user identifier
      schema:
        type: integer
        minimum: 1
    PracticeIdPath:
      name: practice_id
      in: path
      required: true
      description: Numeric practice identifier
      schema:
        type: integer
        minimum: 1
    RoundIdPath:
      name: round_id
      in: path
      required: true
      description: Numeric round identifier
      schema:
        type: integer
        minimum: 1
    EndIdPath:
      name: end_id
      in: path
      required: true
      description: Numeric end identifier
      schema:
        type: integer
        minimum: 1
...
    UserIdFilter:
      name: user_id
      in: query
      required: false
      description: Filter practices by their parent user
      schema:
        type: integer
        minimum: 1
    PracticeIdFilter:
      name: practice_id
      in: query
      required: false
      description: Filter rounds by their parent practice
      schema:
        type: integer
        minimum: 1
    RoundIdFilter:
      name: round_id
      in: query
      required: false
      description: Filter ends by their parent round
      schema:
        type: integer
        minimum: 1
    EndIdFilter:
      name: end_id
      in: query
      required: false
      description: Filter arrows by their parent end
      schema:
        type: integer
        minimum: 1
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Credentials were rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested record was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    LoginInput:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      required:
        - created
        - user
      properties:
        created:
          type: boolean
        password_initialized:
          type: boolean
        user:
          $ref: "#/components/schemas/UserRecord"
    UserRecord:
      type: object
      required:
        - id
        - name
        - created_at
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        created_at:
          type: string
          format: date-time
    UserInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        password:
          type: string
          description: Required on create. Leave empty on update to keep the current password.
    PracticeRecord:
      type: object
      required:
        - id
        - user_id
        - distance_meters
        - target_face_cm
        - notes
        - created_at
      properties:
        id:
          type: integer
          minimum: 1
        user_id:
          type: integer
          minimum: 1
        distance_meters:
          type: integer
          minimum: 1
          default: 50
        target_face_cm:
          type: integer
          minimum: 1
          default: 80
        notes:
          type: string
        created_at:
          type: string
          format: date-time
    PracticeInput:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: integer
          minimum: 1
        distance_meters:
          type: integer
          minimum: 1
          default: 50
        target_face_cm:
          type: integer
          minimum: 1
          default: 80
        notes:
          type: string
          default: ""
    RoundRecord:
      type: object
      required:
        - id
        - practice_id
        - round_order
        - name
        - notes
        - total_score
        - created_at
      properties:
        id:
          type: integer
          minimum: 1
        practice_id:
          type: integer
          minimum: 1
        round_order:
          type: integer
          minimum: 1
        name:
          type: string
        notes:
          type: string
        total_score:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
    RoundInput:
      type: object
      required:
        - practice_id
        - name
      properties:
        practice_id:
          type: integer
          minimum: 1
        name:
          type: string
        notes:
          type: string
          default: ""
    EndRecord:
      type: object
      required:
        - id
        - round_id
        - end_number
        - notes
        - total_score
        - created_at
      properties:
        id:
          type: integer
          minimum: 1
        round_id:
          type: integer
          minimum: 1
        end_number:
          type: integer
          minimum: 1
        notes:
          type: string
        total_score:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
    EndInput:
      type: object
      required:
        - round_id
      properties:
        round_id:
          type: integer
          minimum: 1
        end_number:
          type: integer
          minimum: 1
        notes:
          type: string
          default: ""
...
    ArrowInput:
      type: object
      required:
        - end_id
        - arrow_number
        - score
      properties:
        end_id:
          type: integer
          minimum: 1
        arrow_number:
          type: integer
          minimum: 1
          maximum: 6
        score:
          type: integer
          minimum: 0
          maximum: 10
        score_mark:
          type: string
          enum:
            - ""
            - X
            - M
          default: ""
        x:
          type: number
          format: float
          default: 0
        y:
          type: number
          format: float
          default: 0
        notes:
          type: string
          default: ""